requirements.txt
Django==5.2.13
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-storages==1.14.6
celery==5.5.3
redis==6.4.0
django-filter==25.2
Pillow==11.3.0
psycopg2-binary==2.9.10
django-environ==0.12.0
gunicorn==23.0.0
